backend/requirements.txt
Flask
cloudant
python-dotenv
flask-cors